WIP: Add a drush-based install test.
