+++ devops/phpunit/tests/UpstreamManagementCommandTest.php
@@ -10,41 +10,14 @@ use Symfony\Component\Process\Process;
  */
 class UpstreamManagementCommandTest extends TestCase
 {
-    protected $cleaner;
-    protected $sut;
+    use Fixtures;
 
     public function setUp(): void {
-        $this->cleaner = new Cleaner();
-        $this->cleaner->preventRegistration();
-        $tmpDir = $this->cleaner->tmpdir(sys_get_temp_dir(), 'sut');
-        $this->sut = $tmpDir . DIRECTORY_SEPARATOR . 'sut';
+        $this->initFixtures();
     }
 
     public function tearDown(): void {
-        //$this->cleaner->clean();
-    }
-
-    protected function createSut() {
-        $source = dirname(__DIR__, 3);
-        $filesystem = new Filesystem();
-        foreach (['upstream-configuration', 'web'] as $dir) {
-            $filesystem->mirror($source . DIRECTORY_SEPARATOR . $dir, $this->sut . DIRECTORY_SEPARATOR . $dir);
-        }
-        foreach (scandir($source) as $item) {
-            $fileToCopy = $source . DIRECTORY_SEPARATOR . $item;
-            if (is_file($fileToCopy)) {
-                $filesystem->copy($fileToCopy, $this->sut . DIRECTORY_SEPARATOR . $item);
-            }
-        }
-
-        // Make sure we start out without a lock file in the upstream configuration directory
-        @unlink($this->sut . '/upstream-configuration/composer.lock');
-        @unlink($this->sut . '/composer.lock');
-
-        // Run 'composer update'. This has two important impacts:
-        // 1. The composer.lock file is created, which is necessary for the upstream dependency locking feature to work.
-        // 2. Our preUpdate modifications are applied to the SUT.
-        $this->composer('update');
+        $this->cleanupFixtures();
     }
 
     public function testUpstreamRequire() {
@@ -147,40 +120,5 @@ class UpstreamManagementCommandTest extends TestCase
         $this->assertTrue($process->isSuccessful());
         $this->assertMatchesRegularExpression('#drupal/ctools *4\.#', $output);
         $this->assertDoesNotMatchRegularExpression('#drupal/ctools *4\.0\.2#', $output);
-    }
-
-    public function sutFileContents($file) {
-        return file_get_contents($this->sut . DIRECTORY_SEPARATOR . $file);
-    }
-
-    public function assertSutFileContains($needle, $haystackFile) {
-        $this->assertStringContainsString($needle, $this->sutFileContents($haystackFile));
-    }
-
-    public function assertSutFileNotContains($needle, $haystackFile) {
-        $this->assertStringNotContainsString($needle, $this->sutFileContents($haystackFile));
-    }
-
-    public function assertSutFileDoesNotExist($file) {
-        $this->assertFileDoesNotExist($this->sut . DIRECTORY_SEPARATOR . $file);
-    }
-
-    public function assertSutFileExists($file) {
-        $this->assertFileExists($this->sut . DIRECTORY_SEPARATOR . $file);
-    }
-
-    public function pregReplaceSutFile($regExp, $replace, $file) {
-        $path = $this->sut . DIRECTORY_SEPARATOR . $file;
-        $contents = file_get_contents($path);
-        $contents = preg_replace($regExp, $replace, $contents);
-        file_put_contents($path, $contents);
-    }
-
-    protected function composer(string $command, array $args = []): Process {
-        $cmd = array_merge(['composer', '--working-dir=' . $this->sut, $command], $args);
-        $process = new Process($cmd);
-        $process->run();
-
-        return $process;
     }
 }
